requirements.txt
torch
torchvision
pillow
numpy
fastapi
uvicorn
python-multipart
typing-extensions
beautifulsoup4
yt_dlp
ffmpeg
pytube